requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
idna==2.10
pycodestyle==2.6.0
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.4
requests==2.24.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.25.11
